Currency List and Detail now work
