+++ src/modules/lookup/classes/types/currency/CurrencyUserAction.ts
@@ -1,0 +1,4 @@
+export enum CurrencyUserAction {
+  Refresh = 'refresh',
+  Modify = 'modify'
+}
