+++ src/modules/lookup/classes/types/index.ts
@@ -1,2 +1,4 @@
 export * from './currency/CurrencyField';
+export * from './currency/CurrencyUserAction';
 export * from './systemLimit/SystemLimitField';
+export * from './systemLimit/SystemLimitUserAction';
