+++ src/modules/lookup/classes/types/systemLimit/SystemLimitUserAction.ts
@@ -1,0 +1,4 @@
+export enum SystemLimitUserAction {
+  Refresh = 'refresh',
+  Modify = 'modify'
+}
