+++ src/modules/lookup/components/currency/index.ts
@@ -1,0 +1,4 @@
+export * from './CurrencyListComponent';
+export * from './CurrencyListContainer';
+export * from './CurrencyDetailContainer';
+export * from './CurrencyDetailComponent';
