--- src/modules/lookup/components/list/index.ts
@@ -1,3 +1,0 @@
-export * from './CurrencyList';
-export * from './SystemLimitList';
-export * from './CurrrencyDetail';
