+++ src/modules/lookup/components/systemLimit/index.ts
@@ -1,0 +1,2 @@
+export * from './SystemLimitListComponent';
+export * from './SystemLimitListContainer';
